+++ spec/lib/vcdry/default_spec.rb
@@ -11,5 +11,10 @@ RSpec.describe "VCDry::DSL" do
     it "is expected to allow a proc" do
       expect(component).to have_instance_variable(:@another_string, with_value: "default")
     end
+
+    it "is expected to treat default: nil and optional: true the same" do
+      expect(component).to have_instance_variable(:@integer_1, with_value: nil)
+      expect(component).to have_instance_variable(:@integer_2, with_value: nil)
+    end
   end
 end